+++ example/interpolation.c
@@ -63,11 +63,11 @@ void display(void)
     gluBeginCurve(theNurb);
         gluNurbsCurve(
             theNurb, 
-            spline.n_knots, 
+			(GLint)spline.n_knots,
             spline.knots, 
-            spline.dim, 
+			(GLint)spline.dim,
             spline.ctrlp, 
-            spline.order, 
+			(GLint)spline.order,
             GL_MAP1_VERTEX_3
         );
     gluEndCurve(theNurb);
